vlmo/requirements.txt
pytorch_lightning==1.5.5
transformers==4.8.1
Pillow==8.3.1
tqdm==4.53.0
ipdb==0.13.7
einops==0.3.0
pyarrow==2.0.0
sacred==0.8.2
pandas==1.1.5
timm==0.4.12
torchmetrics==0.7.3
torch==1.8.0
torchvision==0.9.0
fairscale==0.4.0
numpy
scipy
opencv-python